switch to json feed; add interruption check.

+++ trailers.sh
@@ -1,19 +1,27 @@
 #!/usr/bin/env bash
 
 BASENAME="${0##*/}"
+EXPIRY=86400 # 1 day in seconds
 
 test -z "$OUTDIR" && OUTDIR="$HOME/Movies/trailers/"
+test -z "$DATADIR" && DATADIR="${OUTDIR}data/"
 hash jq 2>/dev/null || { echo "$BASENAME requires jq (brew install jq)"; exit 1; }
 hash pup 2>/dev/null || { echo "$BASENAME requires pup (brew install pup)"; exit 1; }
+
+TMP_DIR=$(mktemp -d "$TMPDIR$BASENAME.XXXXXX") || { echo "$BASENAME: can't create the temp dir at: $TMPDIR"; exit 1; }
+TMP_MARKERFILE="$TMP_DIR/marker"
+trap "rm -rf '$TMP_DIR'" EXIT
 
 function __curl () {
   local AE AL UA Ac Rf Cn Ra
   #AE="Accept-Encoding: identity;q=1, *;q=0"
   AL="Accept-Language: en-US,en;q=0.8"
+  #UA="User-Agent: QuickTime/7.6.2"
   UA="User-Agent: Mozilla/5.0 (Macintosh; Intel Mac OS X 10_10_5) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/49.0.2623.112 Safari/537.36"
   Ac="Accept: */*"
   #Rf="Referer: $1"
   Cn="Connection: keep-alive"
+  # TODO: add resuming
   #Ra="Range: bytes=0-"
   curl -H "$AL" -H "$UA" -H "$Ac" -H "$Cn" "$@"
 }
@@ -23,28 +31,44 @@ function __expired () {
   test $(expr $(date +%s) - $(stat -f "%m" "$1")) -gt $2
 }
 
-#FEED_TMPFILE=$(mktemp "$TMPDIR$BASENAME.XXXXXX") || { echo &>2 "cannot make temp file for the feed"; exit 1; }
-FEED_FILE="$TMPDIR$BASENAME.feed.xml"
-#trap "rm '$FEED_TMPFILE'" EXIT
-EXPIRY=86400 # 1 day in seconds
-test -e "$FEED_FILE" && __expired "$FEED_FILE" "$EXPIRY" && mv "$FEED_FILE" "$FEED_FILE.bak" && echo >&2 "$BASENAME: backed expired trailers feed"
+function __fetchFeed () {
+  local FEED FEED_BASENAME FEED_FILE HTTP_RESPONSE
+  FEED="$1"
+  FEED_BASENAME=$(basename "$FEED")
+  FEED_FILE="$TMPDIR$BASENAME.feed.$FEED_BASENAME"
+  test -e "$FEED_FILE" && __expired "$FEED_FILE" "$EXPIRY" && mv "$FEED_FILE" "$FEED_FILE.bak" && echo >&2 "$BASENAME: backed expired trailers feed"
 
-if test -e "$FEED_FILE"; then
-  HTTP_RESPONSE=$(cat "$FEED_FILE")
-else
-  echo >&2 "$BASENAME: fetching trailers feed"
-  FEED=http://images.apple.com/trailers/home/rss/newtrailers.rss
-  #curl -L -s "$FEED" > "$FEED_TMPFILE" || { echo >&2 "$BASENAME: error fetching feed"; exit 1; }
-  HTTP_RESPONSE=$(__curl -L -s "$FEED") || { echo >&2 "$BASENAME: error fetching trailers feed"; exit 1; }
-  test -n "$HTTP_RESPONSE" && echo "$HTTP_RESPONSE" > "$FEED_FILE"
-fi
+  if test -e "$FEED_FILE"; then
+    HTTP_RESPONSE=$(cat "$FEED_FILE")
+  else
+    echo >&2 "$BASENAME: fetching trailers feed"
+    HTTP_RESPONSE=$(__curl -L -s "$FEED") || { echo >&2 "$BASENAME: error fetching trailers feed"; exit 1; }
+    test -n "$HTTP_RESPONSE" && echo "$HTTP_RESPONSE" > "$FEED_FILE"
+  fi
 
-#trap - EXIT
+  echo "$HTTP_RESPONSE"
+}
 
-echo "$HTTP_RESPONSE" | pup 'a + a attr{href}' | while read HREF; do
+#function __processXMLFeed () {
+#  pup 'a + a attr{href}'
+#}
+
+function __processJSONFeed () {
+  jq -r '.[] | "http://trailers.apple.com" + .location'
+}
+
+# NOTE: keep in mind that the XML feed is with smaller number of items
+#HTTP_RESPONSE=$(__fetchFeed "http://images.apple.com/trailers/home/rss/newtrailers.rss")
+#LOCATIONS=$(echo "$HTTP_RESPONSE" | __processXMLFeed)
+HTTP_RESPONSE=$(__fetchFeed "http://trailers.apple.com/trailers/home/feeds/just_added.json")
+LOCATIONS=$(echo "$HTTP_RESPONSE" | __processJSONFeed)
+
+echo "$LOCATIONS" | while read HREF; do
+  test -e "$TMP_MARKERFILE" && { echo -e "\n\n** Interrupted"; exit 2; }
+
   TRAILER_BASENAME=$(basename "$HREF")
   #test "$TRAILER_BASENAME" != "theinvitation" && continue
-  TRAILER_HTMLFILE="$TMPDIR$BASENAME.$TRAILER_BASENAME.html"
+  TRAILER_HTMLFILE="$TMPDIR$BASENAME.page.$TRAILER_BASENAME.html"
   REF="Referer: $HREF"
 
   test -e "$TRAILER_HTMLFILE" && __expired "$TRAILER_HTMLFILE" "$EXPIRY" && mv "$TRAILER_HTMLFILE" "$TRAILER_HTMLFILE.bak" && echo >&2 "$TRAILER_BASENAME: backed expired page"
@@ -53,7 +77,7 @@ echo "$HTTP_RESPONSE" | pup 'a + a attr{href}' | while read HREF; do
     TRAILER_RESPONSE=$(cat "$TRAILER_HTMLFILE")
   else
     echo >&2 "$TRAILER_BASENAME: fetching trailer page"
-    TRAILER_RESPONSE=$(__curl -L "$HREF") || echo >&2 "error fetching trailer page"
+    TRAILER_RESPONSE=$(__curl -L -s "$HREF") || echo >&2 "error fetching trailer page"
     test -n "$TRAILER_RESPONSE" && echo "$TRAILER_RESPONSE" > "$TRAILER_HTMLFILE"
   fi
 
@@ -66,7 +90,7 @@ echo "$HTTP_RESPONSE" | pup 'a + a attr{href}' | while read HREF; do
     else
 
       #TRAILER_JSONFILE="$TMPDIR$BASENAME.$TRAILER_BASENAME.$TRAILER_ID.json"
-      TRAILER_JSONFILE="$OUTDIR$TRAILER_BASENAME.$TRAILER_ID.json"
+      TRAILER_JSONFILE="$DATADIR$TRAILER_BASENAME.$TRAILER_ID.json"
 
       test -e "$TRAILER_JSONFILE" && __expired "$TRAILER_JSONFILE" "$EXPIRY" && mv "$TRAILER_JSONFILE" "$TRAILER_JSONFILE.bak" && "$TRAILER_BASENAME: backed expired json"
 
@@ -78,8 +102,8 @@ echo "$HTTP_RESPONSE" | pup 'a + a attr{href}' | while read HREF; do
         JSON_RESPONSE=$(__curl -L -s -H "$REF" "$TRAILER_JSON") || { echo >&2 "$TRAILER_BASENAME: error fetching trailer json"; JSON_RESPONSE=""; }
         #JSON_RESPONSE=$(echo "$JSON_RESPONSE" | jq '.') || { echo >&2 "$TRAILER_BASENAME: invalid json"; JSON_RESPONSE=""; }
         if test -n "$JSON_RESPONSE"; then
-          if test ! -d "$OUTDIR"; then
-            mkdir -p "$OUTDIR" || { echo >&2 "$BASENAME: couldn't create $OUTDIR"; exit 1; }
+          if test ! -d "$DATADIR"; then
+            mkdir -p "$DATADIR" || { echo >&2 "$BASENAME: couldn't create $DATADIR"; exit 1; }
           fi
           echo "$JSON_RESPONSE" > "$TRAILER_JSONFILE" || echo >&2 "$TRAILER_BASENAME: couldn't store JSON"
         fi
@@ -112,7 +136,9 @@ echo "$HTTP_RESPONSE" | pup 'a + a attr{href}' | while read HREF; do
               if test ! -e "$TRAILER_FILE"; then
                 echo >&2 "$TRAILER_BASENAME: downloading trailer for $MOVIE_TITLE - $TRAILER_FILE_BASENAME"
                 #echo "$URL"
+                touch "$TMP_MARKERFILE"
                 __curl -L -# -H "$REF" "$URL" > "$TRAILER_FILE" || echo "$TRAILER_BASENAME: error downloading the trailer"
+                rm "$TMP_MARKERFILE"
               fi
             done
           fi
